requirements.txt
python-dotenv
langchain_core
langchain
langchain-groq
streamlit
langchain_community
groq
chromadb
streamlit_autorefresh
python-docx
fitz
frontend